+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/models/TripPlan.java
@@ -1,6 +1,7 @@
 package com.parkrangers.parkquest_backend.models;
 
 
+import com.parkrangers.parkquest_backend.models.response.Park;
 import jakarta.persistence.*;
 
 import java.util.ArrayList;
