+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/models/response/Park.java
@@ -1,4 +1,4 @@
-package com.parkrangers.parkquest_backend.models;
+package com.parkrangers.parkquest_backend.models.response;
 
 import jakarta.persistence.*;
