+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/models/response/ParkSearchResponse.java
