+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/repositories/ParkRepository.java
@@ -1,9 +1,11 @@
 package com.parkrangers.parkquest_backend.repositories;
-import com.parkrangers.parkquest_backend.models.Park;
+import com.parkrangers.parkquest_backend.models.response.Park;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface ParkRepository extends JpaRepository<Park, Long> {
-public interface ParkRepository {
+    public interface ParkRepository {
+
+    }
 }
