+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/repositories/TripPlanRepository.java
@@ -1,8 +1,7 @@
 package com.parkrangers.parkquest_backend.repositories;
 
-import com.parkrangers.parkquest_backend.models.Park;
+import com.parkrangers.parkquest_backend.models.response.Park;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
